Dev to master merge (#55)
* issue 17 : ESS충방전량조회 에서 [충전계획] <-> [방전량] (#18)

* issue_21 : BMS 팝업 단위 변경 SOC 현재 kWh -> % (#26)

* 트렐로 이슈정리 2차 (#29)

* 트렐로 이슈정리 2차

회원가입 비밀번호확인 문구 변경, 한전설정화면 감축용량 단위수정, 사용량구성화면 초기 그래프명칭 변경

* 다국어처리 적용 (#30)

* 다극어처리 적용

로그인화면, 군관리메인, 사이트메인, 왼쪽메뉴, 상단, 하단 우선 적용

* 다국어처리 언어권선택 셀렉트박스 추가 (#32)

* 다국어처리 재반영

origin -> upstream:dev로 pr요청시 포함이 되지 않아 따로 반영
원인은 모르겠음

* master, dev 브랜치 동기화 (#36)

* 다국어처리 언어권선택 셀렉트박스 추가 (#33)

* 다국어처리 언어권선택 셀렉트박스 추가 (#32)

* 다국어처리 재반영 (#34)

* 다국어처리 언어권선택 셀렉트박스 추가 (#32)

* 다국어처리 재반영

origin -> upstream:dev로 pr요청시 포함이 되지 않아 따로 반영
원인은 모르겠음

* 자동배포 시작 (#28)

* 자동배포 시작

change whitespace

change ssh options

change persist_to_workspace

add scripts

remove deploy condition

add scripts to remote server

deploy 스크립트 버그 수정

tomcat을 맨처음 중지시키도록 스크립트 변경

sitespeed.io 관련 파일 추가

update circleci

change build option

change test-sitespeed step

성능측정 툴 오류 수정

* master에서만 자동 배포

* issue_24 : ESS수익조회 - 수익도표 에서 충전량 <-> 방전량 (#35)

* 다국어처리 퍼블 적용 및 ess수익조회 화면 수정 (#37)

* 다국어처리 퍼블 수정

* ess수익조회 화면 수정

계절별 경부하,중간부하,최대부하 전력량요금 추가, 명세서 계산로직 수정

* 트렐로 이슈수정(~5/3) 1차 (#38)

* dr실적화현황 화면 실제사용량 합계표시 및 기준부하 단위수정

https://trello.com/c/91t56xL9

* 군관리메인 사이트 목록 순위오류 수정

https://trello.com/c/zCHKNpDR

* 사용량 구성 그래프 값 밸류 수정

https://trello.com/c/E0LsPfSF

* 트렐로 이슈수정(~5/3) 2차 (#39)

* 장치현황 상세보기에서 데이터 없는경우 undefined 문구로 표기되는 오류 수정

https://trello.com/c/7ihuqJTd

* 예측발전량 수집 프로세스 추가 (#14)

* 예측발전량 수집 프로세스 추가

* 예측발전량 pr 1차 수정

* Update config.yml

 VERSION=1.0.1 ->  VERSION=1.0.2 버전 변경

* 요금수익 api 파라미터 단위 조정, ess수익 명세서 계절별 시간대 전력량요금 추가, 스크립트 오류 해결 (#41)

* 요금수익 api 파라미터 단위 조정

* ess수익 명세서 계절별 시간대부하 추가, 스크립트 오류 해결

* ess명세서 충전금액, 방전금액 계산식 수정

* 의미없는 파일 및 디렉토리 삭제 (#42)

* remove target

* ignore target directory

* 트렐로 이슈수정(~5/10) 1차 (#43)

* 사이트 이용자가  장치그룹 편집 가능한 오류 수정

https://trello.com/c/diZGfWqr

* 어드민 관리자 이외의 자기 자신의 레벨 조정이 불가능 하도록 변경

https://trello.com/c/t9dyzVDT

* 자주하는 질문 카테고리 삭제 오류 수정

https://trello.com/c/ueJbk5wu

* 장치 등록시 상태정보 오류 수정

https://trello.com/c/lW05thGz

* 사이트메인 피크전력현황 그래프를 시작시간부터 그리도록 수정 (#44)

* 사이트메인 피크전력현황 그래프를 시작시간부터 그리도록 수정

https://trello.com/c/BVzMGUiR

* 현재시간과 db데이터 기준시간 차이나는 현상 수정 (#47)

* 현재시간과 db데이터 기준시간 차이나는 현상 수정

https://trello.com/c/DGQa6LOS

* Issue 48 - [상황관제]화면의 알림 갯수 색깔 변경 : 빨간색 -> 검은색 (#50)

* [issue_48] (사이트선택 > (메뉴바)상황관제 > 화면) 알림현황 숫자 색 변경 : 빨간색 -> 검정색

* Issue 45 (메뉴와 화면의 띄어쓰기 변경) (#46)

* [한전 계약 및 전력 관리 정보]로 띄어쓰기 변경

* 한전 명세서 계약 전력단위 수정 (#49)

* 한전명세서 계약전력 단위 변경

https://trello.com/c/XsqDyT1T

* 에너지모니터링 조회화면 합계부분을 3자리로 고정 (#52)

* 에너지모니터링 조회화면 합계부분을 3자리로 고정

https://trello.com/c/6VJK6l9b

* 목표사용량 단위 수정 (#53)

* 목표사용량 단위 수정

https://trello.com/c/zur5sG6J

* issue_21 : BMS 팝업 단위 변경 SOC 현재 kWh -> % (#26)

* 다국어처리 적용 (#30)

* 다극어처리 적용

로그인화면, 군관리메인, 사이트메인, 왼쪽메뉴, 상단, 하단 우선 적용

* 다국어처리 언어권선택 셀렉트박스 추가 (#32)

* 다국어처리 재반영

origin -> upstream:dev로 pr요청시 포함이 되지 않아 따로 반영
원인은 모르겠음

* master, dev 브랜치 동기화 (#36)

* 다국어처리 언어권선택 셀렉트박스 추가 (#33)

* 다국어처리 언어권선택 셀렉트박스 추가 (#32)

* 다국어처리 재반영 (#34)

* 다국어처리 언어권선택 셀렉트박스 추가 (#32)

* 다국어처리 재반영

origin -> upstream:dev로 pr요청시 포함이 되지 않아 따로 반영
원인은 모르겠음

* 자동배포 시작 (#28)

* 자동배포 시작

change whitespace

change ssh options

change persist_to_workspace

add scripts

remove deploy condition

add scripts to remote server

deploy 스크립트 버그 수정

tomcat을 맨처음 중지시키도록 스크립트 변경

sitespeed.io 관련 파일 추가

update circleci

change build option

change test-sitespeed step

성능측정 툴 오류 수정

* master에서만 자동 배포

* 다국어처리 퍼블 적용 및 ess수익조회 화면 수정 (#37)

* 다국어처리 퍼블 수정

* ess수익조회 화면 수정

계절별 경부하,중간부하,최대부하 전력량요금 추가, 명세서 계산로직 수정

* 트렐로 이슈수정(~5/3) 1차 (#38)

* dr실적화현황 화면 실제사용량 합계표시 및 기준부하 단위수정

https://trello.com/c/91t56xL9

* 군관리메인 사이트 목록 순위오류 수정

https://trello.com/c/zCHKNpDR

* 사용량 구성 그래프 값 밸류 수정

https://trello.com/c/E0LsPfSF

* 트렐로 이슈수정(~5/3) 2차 (#39)

* 장치현황 상세보기에서 데이터 없는경우 undefined 문구로 표기되는 오류 수정

https://trello.com/c/7ihuqJTd

* 예측발전량 수집 프로세스 추가 (#14)

* 예측발전량 수집 프로세스 추가

* 예측발전량 pr 1차 수정

* 요금수익 api 파라미터 단위 조정, ess수익 명세서 계절별 시간대 전력량요금 추가, 스크립트 오류 해결 (#41)

* 요금수익 api 파라미터 단위 조정

* ess수익 명세서 계절별 시간대부하 추가, 스크립트 오류 해결

* ess명세서 충전금액, 방전금액 계산식 수정

* 의미없는 파일 및 디렉토리 삭제 (#42)

* remove target

* ignore target directory

* 트렐로 이슈수정(~5/10) 1차 (#43)

* 사이트 이용자가  장치그룹 편집 가능한 오류 수정

https://trello.com/c/diZGfWqr

* 어드민 관리자 이외의 자기 자신의 레벨 조정이 불가능 하도록 변경

https://trello.com/c/t9dyzVDT

* 자주하는 질문 카테고리 삭제 오류 수정

https://trello.com/c/ueJbk5wu

* 장치 등록시 상태정보 오류 수정

https://trello.com/c/lW05thGz

* 사이트메인 피크전력현황 그래프를 시작시간부터 그리도록 수정 (#44)

* 사이트메인 피크전력현황 그래프를 시작시간부터 그리도록 수정

https://trello.com/c/BVzMGUiR

* 현재시간과 db데이터 기준시간 차이나는 현상 수정 (#47)

* 현재시간과 db데이터 기준시간 차이나는 현상 수정

https://trello.com/c/DGQa6LOS

* Issue 48 - [상황관제]화면의 알림 갯수 색깔 변경 : 빨간색 -> 검은색 (#50)

* [issue_48] (사이트선택 > (메뉴바)상황관제 > 화면) 알림현황 숫자 색 변경 : 빨간색 -> 검정색

* Issue 45 (메뉴와 화면의 띄어쓰기 변경) (#46)

* [한전 계약 및 전력 관리 정보]로 띄어쓰기 변경

* 한전 명세서 계약 전력단위 수정 (#49)

* 한전명세서 계약전력 단위 변경

https://trello.com/c/XsqDyT1T

* 에너지모니터링 조회화면 합계부분을 3자리로 고정 (#52)

* 에너지모니터링 조회화면 합계부분을 3자리로 고정

https://trello.com/c/6VJK6l9b

* 목표사용량 단위 수정 (#53)

* 목표사용량 단위 수정

https://trello.com/c/zur5sG6J

* resolve conflicts

+++ ewpsp/src/main/java/kr/co/ewp/ewpsp/web/FaqController.java
@@ -230,11 +230,19 @@ public class FaqController {
 		if (userInfo != null) {
 			param.put("modUid", userInfo.get("user_id"));
 		}
-
-		int resultCnt = faqService.deleteFAQCate(param);
-
+		
 		Map<String, Object> resultMap = new HashMap<String, Object>();
-		resultMap.put("resultCnt", resultCnt);
+		
+		param.put("cateIdx", param.get("faqCateIdx"));
+		int cnt = faqService.getFAQListCnt(param);
+		if(cnt > 0) {
+			resultMap.put("msg", "해당 카테고리에 게시물이 있어 삭제할 수 없습니다.");
+		} else {
+			int resultCnt = faqService.deleteFAQCate(param);
+			resultMap.put("msg", "");
+			resultMap.put("resultCnt", resultCnt);
+		}
+		
 		return resultMap;
 	}
